refactor: extract schemas to src/schemas/ package
Move 9 Pydantic model classes out of src/agents/auditor.py and
src/agents/rewriter.py into a dedicated src/schemas/ package.

- src/schemas/lexical.py: SentenceHit, LexicalSummary, LexicalWordReport
- src/schemas/audit.py: SubAgentFinding, SubAgentReport, FlaggedSentence, AuditReport
- src/schemas/rewrite.py: RewriteResult, RewriteReport
- src/schemas/__init__.py: re-exports all 9 classes

All callers (detectors, agents, tests) updated. 44/44 tests pass.
