+++ src/agents/rewriter.py
@@ -1,21 +1,6 @@
 from langfuse import observe
-from pydantic import BaseModel
 
-from src.agents.auditor import FlaggedSentence
-
-
-class RewriteResult(BaseModel):
-    sentence_id: str
-    original: str
-    rewritten: str
-    change_summary: str
-
-
-class RewriteReport(BaseModel):
-    article_id: str
-    model: str
-    rewrites: list[RewriteResult]
-    full_rewritten_text: str
+from src.schemas import FlaggedSentence, RewriteResult, RewriteReport
 
 
 class RewriterAgent:
